Fixing issue #11 for all display types.

+++ 2.7inch_e-paper/raspberrypi/python/epd2in7.py
@@ -244,7 +244,7 @@ class EPD:
             self.send_data(self.lut_wb[count])
 
     def get_frame_buffer(self, image):
-        buf = [0x00] * (self.width * self.height / 8)
+        buf = [0x00] * int(self.width * self.height / 8)
         # Set buffer to value of Python Imaging Library image.
         # Image must be in mode 1.
         image_monocolor = image.convert('1')
@@ -258,19 +258,19 @@ class EPD:
             for x in range(self.width):
                 # Set the bits for the column of pixels at the current position.
                 if pixels[x, y] != 0:
-                    buf[(x + y * self.width) / 8] |= (0x80 >> (x % 8))
+                    buf[int((x + y * self.width) / 8)] |= (0x80 >> (x % 8))
         return buf
 
     def display_frame(self, frame_buffer):
         if (frame_buffer != None):
             self.send_command(DATA_START_TRANSMISSION_1)           
             self.delay_ms(2)
-            for i in range(0, self.width * self.height / 8):
+            for i in range(0, int(self.width * self.height / 8)):
                 self.send_data(0xFF)  
             self.delay_ms(2)                  
             self.send_command(DATA_START_TRANSMISSION_2)
             self.delay_ms(2)
-            for i in range(0, self.width * self.height / 8):
+            for i in range(0, int(self.width * self.height / 8)):
                 self.send_data(frame_buffer[i])  
             self.delay_ms(2)        
             self.send_command(DISPLAY_REFRESH)
